install: Minimize the use of package._bundled
`_bundled` contains the results of `read-package-tree` on a module that contains
bundled dependencies. This is `fetch-package-metadata`'s rather unfortunate approach
to passing this data back to the installer.

We're changing things to remove it from the `package` object ASAP, which in
turn let's us avoid removing it (and then putting it back) in the thing that
writes the package object as `package.json`.