install: Remove failed non-optional deps from tree
This is specifically referring to failures in finding a matching version.

It sounds a little weird, because failing non-optional deps are ordinarily
fatal and would fail out the entire tree, but what this does is ensure that
transitive non-optional deps of an optional dependency get removed from the
tree.  Without this, a failure in one dep of an optional dep would
result in all the other deps of the optional dep still being installed.